proper pyi support after refactor
