+++ src/taipy/gui/builder/page.py
@@ -9,9 +9,4 @@
 # an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-from ..renderers import BlockElementApi, ControlElementApi, _ElementApiGenerator
-
-# separate import for "Page" class so stubgen can properly generate pyi file
-from .page import Page
-
-_ElementApiGenerator().add_default()
+from ..renderers import ClassApi as Page
