+++ src/taipy/gui/renderers/_class_api/__init__.py
@@ -8,3 +8,5 @@
 # Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
 # an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
+from .element_api import BlockElementApi, ControlElementApi, ElementApi
+from .element_api_generator import _ElementApiGenerator
